services/ai-processor/requirements.txt
# Core AI/ML
mediapipe>=0.10.9
ultralytics>=8.0.0
opencv-python>=4.8.0
numpy>=1.24.0

# GPU Acceleration
# cupy-cuda12x>=12.0.0  # Uncomment if using CUDA 12.x
# cupy-cuda11x>=12.0.0  # Uncomment if using CUDA 11.x

# Tracking
filterpy>=1.4.5
scipy>=1.11.0
lap>=0.4.0

# Communication
redis>=5.0.0
grpcio>=1.59.0
grpcio-tools>=1.59.0
asyncio-redis>=0.16.0

# Async & Performance
uvloop>=0.19.0
aiohttp>=3.9.0

# Data Validation & Config
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Logging & Monitoring
loguru>=0.7.0
prometheus-client>=0.18.0

# Utilities
pillow>=10.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.5.0

